packages.yml: echo subpackages to built_packages.txt as well
So that server does not have to look for potential subpackages.  We
check if any of the built (or deleted) packages have subpackages and
in that case pront those to the relevant list as well.

The bintray upload scripts can perhaps be modified a bit now as well,
as it does not need to look for subpackages.