So long Bintray, and thanks for all the fish
In a few days the bintray repos will be readonly, with this no more
updates will be published to the science-packages repo on bintray.

The sponsorship from Bintray has very helpful for us as bandwidth is
not exactly cheap.  From now on it will mainly be up to the community
to serve the repositories through IPFS.